refactor * split up source code into multiple files * doc response object * stub internal methods instead of using a mock server
