+++ packages/annotations/src/index.ts
@@ -1,4 +1,5 @@
 /* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
 
-export * from "./annotations";
+export * from "./search";
+export * from "./util";
